주식비율 소수점 적용

+++ src/main/kotlin/com/redblue/web/company/model/Executive.kt
@@ -41,7 +41,7 @@ data class Executive(
 	@Temporal(TemporalType.DATE)
 	val expiredAt: Date? = null,
 
-	val stockRatio: Int? = 0,
+	val stockRatio: Double? = 0.0,
 
 	val stockCount: Int? = 0,
